rename ReplacePlayerDialog to EditBotsDialog and fix related symbols
