+++ megamek/docs/help/en/PrincessBotDocumentation.html
@@ -217,7 +217,7 @@ Targets list.</SPAN></SPAN></P>
 
 <P>If a player has become disconnected (the word &quot;<I>(ghost)</I>&quot;
     appears next to their name, you can replace that player with a bot.&nbsp;
-    There are two methods for replacing a player. The first is the <i>Replace Player</i> command under the <u>File</u>
+    There are two methods for replacing a player. The first is the <i>Edit Bots</i> command under the <u>File</u>
     menu. This will bring up a version of the Bot Config Dialog along with a list of players that can be replaced. The
     second option is to use the <i>/replacePlayer</i> command in the chat window:</P>
