+++ megamek/i18n/megamek/client/messages.properties
@@ -942,7 +942,7 @@ CommonMenuBar.fileBoardSaveAsImageUnits=Save As Image with Units...
 CommonMenuBar.fileBoardSaveAsImageUnits.tooltip=Save the current board as a PNG image, with all units
 CommonMenuBar.fileGameConnect=Connect...
 CommonMenuBar.fileGameConnectBot=Connect as Bot...
-CommonMenuBar.replacePlayer=Replace Player...
+CommonMenuBar.editBots=Edit Bots...
 CommonMenuBar.fileGameLoad=Load Game...
 CommonMenuBar.fileGameSave=Save Game...
 CommonMenuBar.fileGameSaveServer=Save on Server...
@@ -1845,8 +1845,8 @@ KeyBinds.cmdNames.localLoad=Load Game
 KeyBinds.cmdDesc.localLoad=Start a saved game (only available from MegaMek's main menu)
 KeyBinds.cmdNames.localSave=Save Game 
 KeyBinds.cmdDesc.localSave=Save the current game
-KeyBinds.cmdNames.replacePlayer=Replace Player 
-KeyBinds.cmdDesc.replacePlayer=Replace a faction in a game with a human or bot player
+KeyBinds.cmdNames.editBots=Edit Bots 
+KeyBinds.cmdDesc.editBots=Replace a faction in a game with a human or bot player
 KeyBinds.cmdNames.sensorRange=Visual & Sensor Ranges
 KeyBinds.cmdDesc.sensorRange=Turns the Visual & Sensor Ranges rings on and off
 
@@ -4367,21 +4367,21 @@ BotConfigDialog.saveNewPrompt=Choose a name for the preset
 BotConfigDialog.saveNewTaken=That name already exists.
 BotConfigDialog.previousConfig=Previous Configuration
 
-#Replace Player Dialog
-ReplacePlayersDialog.title=Replace Ghosts Players with Princess Bots and Edit Local Bots
-ReplacePlayersDialog.optionReplace=Replace with Princess
-ReplacePlayersDialog.optionDoNotReplace=Keep as Ghost
-ReplacePlayersDialog.optionNone=No Action
-ReplacePlayersDialog.optionEdit=Edit Bot
-ReplacePlayersDialog.optionKick=Kick Bot
-ReplacePlayersDialog.playerNameHeader=Player
-ReplacePlayersDialog.playerTypeHeader=Type
-ReplacePlayersDialog.configAvailableHeader=Saved Bot Config
-ReplacePlayersDialog.chooseActionHeader=Action
-ReplacePlayersDialog.configAvailable=Restore Saved Config
-ReplacePlayersDialog.config=Bot Settings...
-ReplacePlayersDialog.local=Local
-ReplacePlayersDialog.remote=Remote
+#Edit Bots Dialog
+EditBotsDialog.title=Replace Ghosts Players with Princess Bots and Edit Local Bots
+EditBotsDialog.optionReplace=Replace with Princess
+EditBotsDialog.optionDoNotReplace=Keep as Ghost
+EditBotsDialog.optionNone=No Action
+EditBotsDialog.optionEdit=Edit Bot
+EditBotsDialog.optionKick=Kick Bot
+EditBotsDialog.playerNameHeader=Player
+EditBotsDialog.playerTypeHeader=Type
+EditBotsDialog.configAvailableHeader=Saved Bot Config
+EditBotsDialog.chooseActionHeader=Action
+EditBotsDialog.configAvailable=Restore Saved Config
+EditBotsDialog.config=Bot Settings...
+EditBotsDialog.local=Local
+EditBotsDialog.remote=Remote
 
 #Point Blank Shot Dialog
 StatusBarPhaseDisplay.pointblankShot=Pointblank shot from a hidden unit!
