+++ megamek/i18n/megamek/client/messages_de.properties
@@ -355,7 +355,7 @@ CommonMenuBar.fileGameQuickLoad=Schnellladen
 CommonMenuBar.fileGameSaveServer=Auf dem Server speichern...
 CommonMenuBar.fileUnitsSave=Speichern...
 CommonMenuBar.fileUnitsBrowse=Verfügbare Einheiten...
-CommonMenuBar.replacePlayer=Spieler ersetzen...
+CommonMenuBar.editBots=Spieler ersetzen...
 CommonMenuBar.viewPlayerSettings=Spielereinstellungen
 CommonMenuBar.viewAccessibilityWindow=Zugänglichkeits-Fenster
 CommonMenuBar.viewIncGUIScale=GUI vergrößern
