+++ megamek/i18n/megamek/client/messages_es.properties
@@ -1002,7 +1002,7 @@ CommonMenuBar.fileBoardSaveAsImageUnits=Guardar como Imagen con Unidades...
 CommonMenuBar.fileBoardSaveAsImageUnits.tooltip=Guarde el tablero actual como una imagen PNG, con todas las unidades
 CommonMenuBar.fileGameConnect=Conectar...
 CommonMenuBar.fileGameConnectBot=Conectar un Bot...
-CommonMenuBar.replacePlayer=Sustituir Jugador...
+CommonMenuBar.editBots=Sustituir Jugador...
 CommonMenuBar.fileGameLoad=Cargar Partida...
 CommonMenuBar.fileGameSave=Guardar Partida...
 CommonMenuBar.fileGameSaveServer=Guardar Servidor...
@@ -1770,8 +1770,8 @@ KeyBinds.cmdNames.localLoad=Cargar Partida
 KeyBinds.cmdDesc.localLoad=Iniciar una partida guardada (solo disponible desde el menú principal de MegaMek)
 KeyBinds.cmdNames.localSave=Guardar Partida 
 KeyBinds.cmdDesc.localSave=Guardar la partida actual
-KeyBinds.cmdNames.replacePlayer=Reemplazar Jugador 
-KeyBinds.cmdDesc.replacePlayer=Reemplazar una facción en una partida con un jugador humano o bot
+KeyBinds.cmdNames.editBots=Reemplazar Jugador 
+KeyBinds.cmdDesc.editBots=Reemplazar una facción en una partida con un jugador humano o bot
 
 
 #Key Bindings Overlay
@@ -4210,15 +4210,15 @@ BotConfigDialog.saveNewPrompt=Elija un nombre para el ajuste preestablecido
 BotConfigDialog.saveNewTaken=Ese nombre ya existe.
 BotConfigDialog.previousConfig=Configuración Anterior
 
-#Replace Player Dialog
-ReplacePlayersDialog.title=Reemplazar Jugadores con Princess Bots
-ReplacePlayersDialog.optionReplace=Princess
-ReplacePlayersDialog.optionDoNotReplace=Sin reemplazo
-ReplacePlayersDialog.playerNameHeader=Jugador Fantasma
-ReplacePlayersDialog.configAvailableHeader=Configuración de Princess
-ReplacePlayersDialog.chooseActionHeader=Reemplazar con
-ReplacePlayersDialog.configAvailable=Disponible
-ReplacePlayersDialog.config=Configuración del Bot...
+#Edit Bots Dialog
+EditBotsDialog.title=Reemplazar Jugadores con Princess Bots
+EditBotsDialog.optionReplace=Princess
+EditBotsDialog.optionDoNotReplace=Sin reemplazo
+EditBotsDialog.playerNameHeader=Jugador Fantasma
+EditBotsDialog.configAvailableHeader=Configuración de Princess
+EditBotsDialog.chooseActionHeader=Reemplazar con
+EditBotsDialog.configAvailable=Disponible
+EditBotsDialog.config=Configuración del Bot...
 
 #Point Blank Shot Dialog
 StatusBarPhaseDisplay.pointblankShot=¡Disparo a quemarropa desde una unidad oculta!
