+++ megamek/i18n/megamek/client/messages_ru.properties
@@ -402,7 +402,7 @@ CommonMenuBar.fileBoardSaveAs=Сохранить как...
 CommonMenuBar.fileBoardSaveAsImage=Сохранить как рисунок...
 CommonMenuBar.fileGameConnect=Присоединиться...
 CommonMenuBar.fileGameConnectBot=Присоединиться как бот...
-CommonMenuBar.replacePlayer=Заменить игрока...
+CommonMenuBar.editBots=Заменить игрока...
 CommonMenuBar.fileGameNew=Новая
 CommonMenuBar.fileGameOpen=Открыть...
 CommonMenuBar.fileGameSave=Сохранить локально...
