+++ megamek/src/megamek/client/ui/swing/RandomArmyDialog.java
@@ -58,10 +58,10 @@ public class RandomArmyDialog extends JDialog implements ActionListener, TreeSel
     private static final int TAB_RAT_GENERATOR     = 2;
     private static final int TAB_FORMATION_BUILDER = 3;
     private static final int TAB_FORCE_GENERATOR   = 4;
-    
+
     private static final String CARD_PREVIEW = "card_preview";
     private static final String CARD_FORCE_TREE = "card_force_tree";
-    
+
     private ClientGUI m_clientgui;
     private Client m_client;
     AdvancedSearchDialog2 asd;
@@ -154,7 +154,7 @@ public class RandomArmyDialog extends JDialog implements ActionListener, TreeSel
             .getString("RandomArmyDialog.Pad"));
     private JCheckBox m_chkCanon = new JCheckBox(Messages
             .getString("RandomArmyDialog.Canon"));
-    
+
     private RandomUnitGenerator rug;
     private UnitTable generatedRAT;
 
@@ -178,7 +178,7 @@ public class RandomArmyDialog extends JDialog implements ActionListener, TreeSel
         rug = RandomUnitGenerator.getInstance();
         rug.registerListener(this);
         if (rug.isInitialized()) {
-            m_ratStatus = new JLabel(Messages.getString("RandomArmyDialog.ratStatusDoneLoading"));            
+            m_ratStatus = new JLabel(Messages.getString("RandomArmyDialog.ratStatusDoneLoading"));
         } else {
             m_ratStatus = new JLabel(Messages.getString("RandomArmyDialog.ratStatusLoading"));
         }
@@ -208,7 +208,7 @@ public class RandomArmyDialog extends JDialog implements ActionListener, TreeSel
                 this.m_bRandomSkills.setEnabled(true);
             }
         });
-        
+
         m_pRightPane.add(m_pPreview, CARD_PREVIEW);
         m_pRightPane.add(m_pForceGen.getRightPanel(), CARD_FORCE_TREE);
         m_pRightPane.setMinimumSize(new Dimension(0,0));
@@ -223,13 +223,13 @@ public class RandomArmyDialog extends JDialog implements ActionListener, TreeSel
         add(m_pSplit, BorderLayout.CENTER);
         validate();
         setLocationRelativeTo(cl.frame);
-        
+
         m_pSplit.setDividerLocation(GUIP.getRndArmySplitPos());
         setSize(GUIP.getRndArmySizeWidth(), GUIP.getRndArmySizeHeight());
         setLocation(GUIP.getRndArmyPosX(), GUIP.getRndArmyPosY());
 
         adaptToGUIScale();
-        
+
         m_client.getGame().addGameListener(gameListener);
         addWindowListener(windowListener);
     }
@@ -644,16 +644,16 @@ public class RandomArmyDialog extends JDialog implements ActionListener, TreeSel
                 Client c = null;
                 if (m_chPlayer.getSelectedIndex() > 0) {
                     String name = (String) m_chPlayer.getSelectedItem();
-                    c = m_clientgui.getBots().get(name);
+                    c = m_clientgui.getLocalBots().get(name);
                 }
                 if (c == null) {
                     c = m_client;
                 }
                 for (MechSummary ms : armyModel.getAllUnits()) {
                     try {
-                        Entity e = new MechFileParser(ms.getSourceFile(), 
+                        Entity e = new MechFileParser(ms.getSourceFile(),
                                 ms.getEntryName()).getEntity();
-      
+
                         autoSetSkillsAndName(e);
                         e.setOwner(c.getLocalPlayer());
                         if (!c.getGame().getPhase().isLounge()) {
@@ -679,7 +679,7 @@ public class RandomArmyDialog extends JDialog implements ActionListener, TreeSel
                 m_lUnitsBVTotal.setText(msg_bvtotal + "0");
                 m_lArmyBVTotal.setText(msg_bvtotal + "0");
             }
-            
+
             // Save preferences
             GUIP.setRATBVMin(m_tBVmin.getText());
             GUIP.setRATBVMax(m_tBVmax.getText());
@@ -696,7 +696,7 @@ public class RandomArmyDialog extends JDialog implements ActionListener, TreeSel
             } else {
                 GUIP.setRATSelectedRAT("");
             }
-            
+
             setVisible(false);
         } else if (ev.getSource().equals(m_bClear)) {
             armyModel.clearData();
@@ -760,7 +760,7 @@ public class RandomArmyDialog extends JDialog implements ActionListener, TreeSel
                             EnumSet.noneOf(MissionRole.class), 0, fRec));
                     List<Integer> numUnits = new ArrayList<>();
                     numUnits.add(m_pFormationOptions.getNumUnits());
-                    
+
                     if (m_pFormationOptions.getIntegerOption("numOtherUnits") > 0) {
                         if (m_pFormationOptions.getIntegerOption("otherUnitType") >= 0) {
                             params.add(new UnitTable.Parameters(fRec,
@@ -782,7 +782,7 @@ public class RandomArmyDialog extends JDialog implements ActionListener, TreeSel
                             //BA do not count for formation rules; add as a separate formation
                         }
                     }
-                    
+
                     if (ft != null) {
                         unitList.addAll(ft.generateFormation(params,
                                 numUnits, m_pFormationOptions.getIntegerOption("network"), false));
@@ -926,14 +926,14 @@ public class RandomArmyDialog extends JDialog implements ActionListener, TreeSel
             GUIP.setRndArmySplitPos(m_pSplit.getDividerLocation());
         }
     };
-    
+
     private void updatePlayerChoice() {
         String lastChoice = (String) m_chPlayer.getSelectedItem();
         String clientName = m_clientgui.getClient().getName();
         m_chPlayer.removeAllItems();
         m_chPlayer.setEnabled(true);
         m_chPlayer.addItem(clientName);
-        for (Client client : m_clientgui.getBots().values()) {
+        for (Client client : m_clientgui.getLocalBots().values()) {
             Player player = m_client.getGame().getPlayer(client.getLocalPlayerNumber());
 
             if (!player.isObserver()) {
@@ -985,13 +985,13 @@ public class RandomArmyDialog extends JDialog implements ActionListener, TreeSel
         Iterator<String> rats = rug.getRatList();
         if (null == rats) {
             return;
-        }  
+        }
 
         RatTreeNode ratTree = rug.getRatTree();
         DefaultMutableTreeNode root = new DefaultMutableTreeNode(ratTree.name);
         createRatTreeNodes(root, ratTree);
         m_treeRAT.setModel(new DefaultTreeModel(root));
-        
+
         String selectedRATPath = GUIP.getRATSelectedRAT();
         if (!selectedRATPath.isBlank()) {
             String[] nodes = selectedRATPath.replace('[', ' ')
@@ -1000,7 +1000,7 @@ public class RandomArmyDialog extends JDialog implements ActionListener, TreeSel
             m_treeRAT.setSelectionPath(path);
         }
     }
-    
+
     private void updateRATYear() {
         GameOptions gameOptions = m_client.getGame().getOptions();
         int gameYear = gameOptions.intOption("year");
@@ -1167,7 +1167,7 @@ public class RandomArmyDialog extends JDialog implements ActionListener, TreeSel
             }
         }
     };
-    
+
     /**
      * A table model for displaying units
      */
@@ -1275,7 +1275,7 @@ public class RandomArmyDialog extends JDialog implements ActionListener, TreeSel
             return data;
         }
     }
-    
+
     /**
      * A table model for displaying a generated RAT
      */
@@ -1303,7 +1303,7 @@ public class RandomArmyDialog extends JDialog implements ActionListener, TreeSel
         public int getColumnCount() {
             return N_COL;
         }
-        
+
         public int getPreferredWidth(int col) {
             switch (col) {
                 case COL_WEIGHT:
@@ -1320,7 +1320,7 @@ public class RandomArmyDialog extends JDialog implements ActionListener, TreeSel
                     return 0;
             }
         }
-        
+
         @Override
         public String getColumnName(int column) {
             switch (column) {
