+++ megamek/src/megamek/client/ui/swing/RandomNameDialog.java
@@ -104,7 +104,7 @@ public class RandomNameDialog extends JDialog implements ActionListener {
         chPlayer.setEnabled(true);
         chPlayer.addItem(clientName);
 
-        for (Iterator<Client> i = clientgui.getBots().values().iterator(); i.hasNext();) {
+        for (Iterator<Client> i = clientgui.getLocalBots().values().iterator(); i.hasNext();) {
             chPlayer.addItem(i.next().getName());
         }
         if (chPlayer.getItemCount() == 1) {
@@ -143,7 +143,7 @@ public class RandomNameDialog extends JDialog implements ActionListener {
             Client c = null;
             if (chPlayer.getSelectedIndex() > 0) {
                 String name = (String) chPlayer.getSelectedItem();
-                c = clientgui.getBots().get(name);
+                c = clientgui.getLocalBots().get(name);
             }
             if (c == null) {
                 c = client;
