+++ megamek/src/megamek/client/ui/swing/dialog/MegaMekUnitSelectorDialog.java
@@ -103,7 +103,7 @@ public class MegaMekUnitSelectorDialog extends AbstractUnitSelectorDialog {
             String name = (String) comboPlayer.getSelectedItem();
 
             if (comboPlayer.getSelectedIndex() > 0) {
-                client = clientGUI.getBots().get(name);
+                client = clientGUI.getLocalBots().get(name);
             }
 
             if (client == null) {
@@ -145,7 +145,7 @@ public class MegaMekUnitSelectorDialog extends AbstractUnitSelectorDialog {
         comboPlayer.removeAllItems();
         comboPlayer.setEnabled(true);
         comboPlayer.addItem(clientName);
-        for (Client client : clientGUI.getBots().values()) {
+        for (Client client : clientGUI.getLocalBots().values()) {
             comboPlayer.addItem(client.getName());
         }
         if (comboPlayer.getItemCount() == 1) {
