+++ megamek/src/megamek/client/ui/swing/lobby/LobbyActions.java
@@ -15,7 +15,7 @@
  *
  * You should have received a copy of the GNU General Public License
  * along with MegaMek. If not, see <http://www.gnu.org/licenses/>.
- */ 
+ */
 package megamek.client.ui.swing.lobby;
 
 import megamek.client.Client;
@@ -66,7 +66,7 @@ public class LobbyActions {
         lobby = cl;
     }
 
-    /** Sets a deployment round for the given entities. Sends an update to the server. */ 
+    /** Sets a deployment round for the given entities. Sends an update to the server. */
     void applyDeployment(Collection<Entity> entities, int newRound) {
         if (!validateUpdate(entities)) {
             return;
@@ -138,9 +138,9 @@ public class LobbyActions {
         }
         sendUpdates(updateCandidates);
     }
-    
+
     /**
-     * Attaches the given force as a subforce to the given new parent. 
+     * Attaches the given force as a subforce to the given new parent.
      * Does NOT work for newParentId == NO_FORCE. Use promoteForce to do this.
      * Does not allow attaching a force to one of its own subforces.
      */
@@ -150,13 +150,13 @@ public class LobbyActions {
                 || (forceId == newParentId)) {
             return;
         }
-        
+
         Force force = forces.getForce(forceId);
         Force newParent = forces.getForce(newParentId);
         List<Force> subForces = forces.getFullSubForces(force);
         Player owner = forces.getOwner(force);
         Player newParentOwner = forces.getOwner(newParent);
-            
+
         if (owner.isEnemyOf(newParentOwner)) {
             LobbyErrors.showOnlyTeam(frame());
             return;
@@ -172,10 +172,10 @@ public class LobbyActions {
         var forceList = new ArrayList<>(List.of(force));
         client().sendForceParent(forceList, newParentId);
     }
-    
+
 
     /**
-     * Makes the given forces top-level, detaching them from any former parents. 
+     * Makes the given forces top-level, detaching them from any former parents.
      */
     void forcePromote(Collection<Integer> forceIds) {
         var forces = game().getForces();
@@ -204,10 +204,10 @@ public class LobbyActions {
         med.setVisible(true);
         sendUpdates(entities);
     }
-    
-    /** 
+
+    /**
      * Moves a force or entity within another force by one position. If up is true,
-     * moves upward, otherwise downward. 
+     * moves upward, otherwise downward.
      */
     void forceMove(Collection<Force> forceList, Collection<Entity> entityList, boolean up) {
         // May only move a single force or a single entity
@@ -223,7 +223,7 @@ public class LobbyActions {
             return;
         }
         var forces = game().getForces();
-        var changedForce = new HashSet<Force>(); 
+        var changedForce = new HashSet<Force>();
         if (up) {
             if (!forceList.isEmpty()) {
                 changedForce.addAll(forces.moveUp(CollectionUtil.anyOneElement(forceList)));
@@ -242,10 +242,10 @@ public class LobbyActions {
             client().sendUpdateForce(changedForce);
         }
     }
-    
-    /** 
-     * Displays a CamoChooser to choose an individual camo for the given entities. 
-     * The camo will only be applied to units configurable by the local player, 
+
+    /**
+     * Displays a CamoChooser to choose an individual camo for the given entities.
+     * The camo will only be applied to units configurable by the local player,
      * i.e. his own units or those of his bots.
      */
     public void individualCamo(Collection<Entity> entities) {
@@ -288,10 +288,10 @@ public class LobbyActions {
         String ownerName = randomSelected.getOwner().getName();
         int ownerId = randomSelected.getOwner().getId();
 
-        boolean editable = client().bots.get(ownerName) != null;
+        boolean editable = client().localBots.get(ownerName) != null;
         Client client;
         if (editable) {
-            client = client().bots.get(ownerName);
+            client = client().localBots.get(ownerName);
         } else {
             editable |= ownerId == localPlayer().getId();
             client = client();
@@ -300,7 +300,7 @@ public class LobbyActions {
         CustomMechDialog cmd = new CustomMechDialog(lobby.getClientgui(), client, new ArrayList<>(entities), editable);
         cmd.setSize(new Dimension(GUIPreferences.getInstance().getCustomUnitWidth(),
                 GUIPreferences.getInstance().getCustomUnitHeight()));
-        cmd.setTitle(Messages.getString("ChatLounge.CustomizeUnits")); 
+        cmd.setTitle(Messages.getString("ChatLounge.CustomizeUnits"));
         cmd.setVisible(true);
         GUIPreferences.getInstance().setCustomUnitHeight(cmd.getSize().height);
         GUIPreferences.getInstance().setCustomUnitWidth(cmd.getSize().width);
@@ -347,10 +347,10 @@ public class LobbyActions {
         if (!validateUpdate(Arrays.asList(entity))) {
             return;
         }
-        boolean editable = client().bots.get(entity.getOwner().getName()) != null;
+        boolean editable = client().localBots.get(entity.getOwner().getName()) != null;
         Client c;
         if (editable) {
-            c = client().bots.get(entity.getOwner().getName());
+            c = client().localBots.get(entity.getOwner().getName());
         } else {
             editable |= entity.getOwnerId() == localPlayer().getId();
             c = client();
@@ -429,9 +429,9 @@ public class LobbyActions {
         }
     }
 
-    /** 
+    /**
      * Sets random skills for the given entities, as far as they can
-     * be configured by the local player. 
+     * be configured by the local player.
      */
     void setRandomSkills(Collection<Entity> entities) {
         if (!validateUpdate(entities)) {
@@ -444,9 +444,9 @@ public class LobbyActions {
         sendUpdates(entities);
     }
 
-    /** 
+    /**
      * Sets random names for the given entities' pilots, as far as they can
-     * be configured by the local player. 
+     * be configured by the local player.
      */
     void setRandomNames(Collection<Entity> entities) {
         if (!validateUpdate(entities)) {
@@ -462,9 +462,9 @@ public class LobbyActions {
         sendUpdates(entities);
     }
 
-    /** 
+    /**
      * Sets random callsigns for the given entities' pilots, as far as they can
-     * be configured by the local player. 
+     * be configured by the local player.
      */
     void setRandomCallsigns(Collection<Entity> entities) {
         if (!validateUpdate(entities)) {
@@ -477,9 +477,9 @@ public class LobbyActions {
         }
         sendUpdates(entities);
     }
-    
+
     /**
-     * Asks for a name and creates a new top-level force of that name. 
+     * Asks for a name and creates a new top-level force of that name.
      */
     void forceCreateEmpty() {
         // Ask for a name
@@ -489,10 +489,10 @@ public class LobbyActions {
         }
         client().sendAddForce(Force.createToplevelForce(name, localPlayer()), new ArrayList<>());
     }
-    
+
     /**
-     * Asks for a name and creates a new top-level force of that name with the 
-     * selected entities in it. 
+     * Asks for a name and creates a new top-level force of that name with the
+     * selected entities in it.
      */
     void forceCreateFrom(Collection<Entity> entities) {
         if (!validateUpdate(entities)) {
@@ -509,10 +509,10 @@ public class LobbyActions {
         }
         client().sendAddForce(Force.createToplevelForce(name, CollectionUtil.anyOneElement(entities).getOwner()), entities);
     }
-    
+
     /**
      * Asks for a name and creates a new subforce of that name for the force given
-     * as the parentId. 
+     * as the parentId.
      */
     void forceCreateSub(int parentId) {
         // Ask for a name
@@ -522,7 +522,7 @@ public class LobbyActions {
         }
         client().sendAddForce(Force.createSubforce(name, game().getForces().getForce(parentId)), new ArrayList<>());
     }
-    
+
     /**
      * Toggles burst MG fire for the given entities to the state given as burstOn
      */
@@ -541,17 +541,17 @@ public class LobbyActions {
         }
         sendUpdates(updateCandidates);
     }
-    
+
     /** Adds the given entities as strategic targets for the given local bot. */
     void setPrioTarget(String botName, Collection<Entity> entities) {
-        Map<String, Client> bots = lobby.getClientgui().getBots();
+        Map<String, Client> bots = lobby.getClientgui().getLocalBots();
         if (!bots.containsKey(botName) || !(bots.get(botName) instanceof Princess)) {
             return;
         }
         BehaviorSettings behavior = ((Princess) bots.get(botName)).getBehaviorSettings();
         entities.forEach(e -> behavior.addPriorityUnit(e.getId()));
     }
-    
+
     /**
      * Toggles hot loading LRMs for the given entities to the state given as hotLoadOn
      */
@@ -561,7 +561,7 @@ public class LobbyActions {
         }
         Set<Entity> updateCandidates = new HashSet<>();
         for (Entity entity: entities) {
-            for (Mounted m: entity.getAmmo()) { 
+            for (Mounted m: entity.getAmmo()) {
                 // setHotLoad checks the Ammo to see if it can be hotloaded
                 m.setHotLoad(hotLoadOn);
                 // TODO: The following should ideally be part of setHotLoad in Mounted
@@ -575,7 +575,7 @@ public class LobbyActions {
         }
         sendUpdates(updateCandidates);
     }
-    
+
     public void load(Collection<Entity> selEntities, String info) {
         StringTokenizer stLoad = new StringTokenizer(info, ":");
         int loaderId = Integer.parseInt(stLoad.nextToken());
@@ -587,7 +587,7 @@ public class LobbyActions {
         if (entities.isEmpty()) {
             return;
         }
-        
+
         // If a unit of the selected units is currently loaded onto another, 2nd unit of the selected
         // units, do not continue. The player should unload units first. This would require
         // a server update offloading that second unit AND embarking it. Currently not possible
@@ -599,7 +599,7 @@ public class LobbyActions {
             LobbyErrors.showNoDualLoad(frame());
             return;
         }
-        
+
         boolean loadRear = false;
         if (stLoad.hasMoreTokens()) {
             loadRear = Boolean.parseBoolean(stLoad.nextToken());
@@ -607,7 +607,7 @@ public class LobbyActions {
 
         StringBuilder errorMsg = new StringBuilder();
         if (!LobbyUtility.validateLobbyLoad(entities, loader, bayNumber, loadRear, errorMsg)) {
-            JOptionPane.showMessageDialog(frame(), errorMsg.toString(), 
+            JOptionPane.showMessageDialog(frame(), errorMsg.toString(),
                     Messages.getString("LoadingBay.error"), JOptionPane.ERROR_MESSAGE);
             return;
         }
@@ -624,14 +624,14 @@ public class LobbyActions {
             lobby.loadOnto(e, loaderId, bayNumber);
         }
     }
-    
+
     /** Asks for a new name for the provided forceId and applies it. */
     void forceRename(int forceId) {
         Forces forces = game().getForces();
         if (!forces.contains(forceId)) {
             return;
         }
-        Force force = forces.getForce(forceId); 
+        Force force = forces.getForce(forceId);
         if (!isEditable(force)) {
             LobbyErrors.showCannotConfigEnemies(frame());
             return;
@@ -645,13 +645,13 @@ public class LobbyActions {
         var forceList = new ArrayList<>(List.of(force)); // must be mutable
         client().sendUpdateForce(forceList);
     }
-    
+
     /**
-     * Deletes the given forces and entities. Asks for confirmation if confirm is true. 
+     * Deletes the given forces and entities. Asks for confirmation if confirm is true.
      */
     void delete(Collection<Force> foDelete, Collection<Entity> enDelete, boolean confirm) {
         Forces forces = game().getForces();
-        // Remove redundant forces = subforces of other forces in the list 
+        // Remove redundant forces = subforces of other forces in the list
         Set<Force> allSubForces = new HashSet<>();
         foDelete.forEach(f -> allSubForces.addAll(forces.getFullSubForces(f)));
         foDelete.removeIf(allSubForces::contains);
@@ -661,7 +661,7 @@ public class LobbyActions {
         foDelete.stream().map(forces::getFullEntities).map(ForceAssignable::filterToEntityList).forEach(inForces::addAll);
         enDelete.removeIf(inForces::contains);
         Set<Entity> finalEnDelete = new HashSet<>(enDelete);
-        
+
         if (!enDelete.isEmpty() && !validateUpdate(finalEnDelete)) {
             return;
         }
@@ -686,17 +686,17 @@ public class LobbyActions {
                 return;
             }
         }
-        
+
         // Send a command to remove the forceless entities
         Set<Client> senders = finalEnDelete.stream().map(this::correctSender).collect(toSet());
         for (Client sender: senders) {
-            // Gather the entities for this sending client; 
+            // Gather the entities for this sending client;
             // Serialization doesn't like the toList() result, therefore the new ArrayList
             List<Integer> ids = new ArrayList<>(finalEnDelete.stream()
                     .filter(e -> correctSender(e).equals(sender)).map(Entity::getId).collect(toList()));
             sender.sendDeleteEntities(ids);
         }
-        
+
         // Send a command to remove the forces (with entities)
         senders = finalFoDelete.stream().map(this::correctSender).collect(toSet());
         for (Client sender: senders) {
@@ -706,11 +706,11 @@ public class LobbyActions {
             sender.sendDeleteForces(foList);
         }
     }
-    
+
     /**
      * Removes the given entities from their force(s), making them force-less.
      * Entities must have a single owner and be editable (local units or local bot's units)
-     * (Having multiple owners makes sending updates correctly for one's own bots difficult) 
+     * (Having multiple owners makes sending updates correctly for one's own bots difficult)
      */
     void forceRemoveEntity(Collection<Entity> entities) {
         if (!validateUpdate(entities)) {
@@ -718,9 +718,9 @@ public class LobbyActions {
         }
         client().sendAddEntitiesToForce(entities, Force.NO_FORCE);
     }
-    
+
     /**
-     * Swaps pilots between the given entity 
+     * Swaps pilots between the given entity
      * and another entity of the given id
      */
     void swapPilots(Collection<Entity> entities, int targetId) {
@@ -741,12 +741,12 @@ public class LobbyActions {
         selected.setCrew(temp);
         sendUpdates(Arrays.asList(target, selected));
     }
-    
-    /** 
+
+    /**
      * Disconnects the passed entities from their C3 network, if any.
      * Due to the way C3 networks are represented in Entity, units
      * cannot disconnect from a C3 network with an id that is the
-     * entity's own id. 
+     * entity's own id.
      */
     void c3DisconnectFromNetwork(Collection<Entity> entities) {
         if (!validateUpdate(entities)) {
@@ -755,10 +755,10 @@ public class LobbyActions {
         Set<Entity> updateCandidates = performDisconnect(entities);
         sendUpdates(updateCandidates);
     }
-    
-    /** 
-     * Performs a disconnect from C3 networks for the given entities without sending an update. 
-     * Returns a set of all affected units. 
+
+    /**
+     * Performs a disconnect from C3 networks for the given entities without sending an update.
+     * Returns a set of all affected units.
      */
     private HashSet<Entity> performDisconnect(Collection<Entity> entities) {
         HashSet<Entity> updateCandidates = new HashSet<>();
@@ -780,7 +780,7 @@ public class LobbyActions {
         }
         return updateCandidates;
     }
-    
+
     /**  Sets the entities' C3M to act as a Company Master. */
     void c3SetCompanyMaster(Collection<Entity> entities) {
         if (!validateUpdate(entities)) {
@@ -793,7 +793,7 @@ public class LobbyActions {
         entities.forEach(e -> e.setC3Master(e.getId(), true));
         sendUpdates(entities);
     }
-    
+
     /**  Sets the entities' C3M to act as a Lance Master (aka normal mode). */
     void c3SetLanceMaster(Collection<Entity> entities) {
         if (!validateUpdate(entities)) {
@@ -806,8 +806,8 @@ public class LobbyActions {
         entities.forEach(e -> e.setC3Master(-1, true));
         sendUpdates(entities);
     }
-    
-    /** 
+
+    /**
      * Connects the passed entities to a nonhierarchic C3 (NC3, C3i or Nova CEWS)
      * identified by masterID.
      */
@@ -837,7 +837,7 @@ public class LobbyActions {
         sendUpdates(entities);
     }
 
-    /** 
+    /**
      * Connects the passed entities to a standard C3M
      * identified by masterID.
      */
@@ -873,8 +873,8 @@ public class LobbyActions {
         entities.forEach(e -> e.setC3Master(master, true));
         sendUpdates(updateCandidates);
     }
-    
-    /** 
+
+    /**
      * Change the given entities' controller to the player with ID newOwnerId.
      * If the given forceList is not empty, an error message will be shown.
      */
@@ -890,7 +890,7 @@ public class LobbyActions {
         }
         client().sendChangeOwner(entities, newOwnerId);
     }
-    
+
     /** Change the team of a controlled player (the local player or one of his bots). */
     void changeTeam(Collection<Player> players, int team) {
         var toSend = new HashSet<Player>();
@@ -900,7 +900,7 @@ public class LobbyActions {
             .forEach(toSend::add);
         client().sendChangeTeam(toSend, team);
     }
-    
+
     /**
      * Add the entities to the force if admissible (the entities must all be editable
      * by the local player and be allied to the force's owner.
@@ -917,14 +917,14 @@ public class LobbyActions {
         }
         client().sendAddEntitiesToForce(entities, forceId);
     }
-    
-    /** 
-     * Changes the owner of the given forces to a different player without 
-     * affecting force structure. 
+
+    /**
+     * Changes the owner of the given forces to a different player without
+     * affecting force structure.
      * When assigning the force only to an enemy, it would dislodge that force
      * from its parent and dislodge all units from it and leave it an empty
      * force for the enemy. That seems useless. Therefore this is restricted
-     * to only assign to team members of the former owner. 
+     * to only assign to team members of the former owner.
      */
     void forceAssignOnly(Collection<Force> forceList, int newOwnerId) {
         Player newOwner = game().getPlayer(newOwnerId);
@@ -946,8 +946,8 @@ public class LobbyActions {
         }
         client().sendUpdateForce(changedForces);
     }
-    
-    /** 
+
+    /**
      * Changes the owner of the given forces to a different player together with
      * all subforces and units.
      */
@@ -962,7 +962,7 @@ public class LobbyActions {
         }
         client().sendAssignForceFull(forceList, newOwnerId);
     }
-    
+
     void unloadFromBay(Collection<Entity> entities, int bayId) {
         if (entities.size() != 1) {
             LobbyErrors.showSingleUnit(frame(), "offload from bays");
@@ -983,7 +983,7 @@ public class LobbyActions {
         }
         sendUpdates(updateCandidates);
     }
-    
+
     /**
      * Creates a fighter squadron from the given list of entities.
      * Checks if all entities are fighters and if the number of entities
@@ -1002,35 +1002,35 @@ public class LobbyActions {
             return;
         }
         boolean largeSquadrons = game().getOptions().booleanOption(OptionsConstants.ADVAERORULES_ALLOW_LARGE_SQUADRONS);
-        if ((!largeSquadrons && entities.size() > FighterSquadron.MAX_SIZE) 
+        if ((!largeSquadrons && entities.size() > FighterSquadron.MAX_SIZE)
                 || entities.size() > FighterSquadron.ALTERNATE_MAX_SIZE) {
             LobbyErrors.showSquadronTooMany(frame());
         }
-        
+
         // Ask for a squadron name
         String name = JOptionPane.showInputDialog(frame(), "Choose a squadron designation");
         if ((name == null) || name.isBlank()) {
             return;
         }
-        
+
         // Now, actually create the squadron
         FighterSquadron fs = new FighterSquadron(name);
         fs.setOwner(createSquadronOwner(entities));
         List<Integer> fighterIds = new ArrayList<>(entities.stream().map(Entity::getId).collect(toList()));
         correctSender(fs).sendAddSquadron(fs, fighterIds);
     }
-    
-    /** 
+
+    /**
      * Returns a likely owner client; if any of the fighter belongs to the local
      * player, returns the local player. If not, returns a local bot if any of the
-     * fighters belongs to that; finally, returns the owner of a random one of the 
+     * fighters belongs to that; finally, returns the owner of a random one of the
      * fighters.
      */
     private Player createSquadronOwner(Collection<Entity> entities) {
         if (entities.stream().anyMatch(e -> e.getOwner().equals(localPlayer()))) {
             return localPlayer();
         } else {
-            for (Entry<String, Client> en: client().bots.entrySet()) {
+            for (Entry<String, Client> en: client().localBots.entrySet()) {
                 Player bot = en.getValue().getLocalPlayer();
                 if (entities.stream().anyMatch(e -> e.getOwner().equals(bot))) {
                     return en.getValue().getLocalPlayer();
@@ -1073,17 +1073,17 @@ public class LobbyActions {
         return true;
     }
 
-    /** 
-     * Sends the entities in the given Collection to the Server. 
+    /**
+     * Sends the entities in the given Collection to the Server.
      * Sends only those that can be edited, i.e. the player's own
      * or his bots' units. Will separate the units into update
-     * packets for the local player and any local bots so that the 
+     * packets for the local player and any local bots so that the
      * server accepts all changes (as the server does not know of
      * local bots and rejects updates that are not for the sending client
-     * or its teammates. 
+     * or its teammates.
      */
     void sendUpdates(Collection<Entity> entities) {
-        // Gather the necessary sending clients; this list may contain null if some units 
+        // Gather the necessary sending clients; this list may contain null if some units
         // cannot be affected at all, i.e. are enemies to localplayer and all his bots
         List<Client> senders = entities.stream().map(this::correctSender).distinct().collect(toList());
         for (Client sender: senders) {
@@ -1093,47 +1093,47 @@ public class LobbyActions {
             sender.sendUpdateEntity(new ArrayList<>(entities.stream().filter(e -> correctSender(e).equals(sender)).collect(toList())));
         }
     }
-    
-    /** 
-     * Sends the entities and forces in the given Collections to the Server. 
+
+    /**
+     * Sends the entities and forces in the given Collections to the Server.
      * Sends only those that can be edited, i.e. the player's own
      * or his bots' units. Will separate the units into update
-     * packets for the local player and any local bots so that the 
+     * packets for the local player and any local bots so that the
      * server accepts all changes (as the server does not know of
      * local bots and rejects updates that are not for the sending client
-     * or its teammates. 
+     * or its teammates.
      */
     void sendUpdates(Collection<Entity> changedEntities, Collection<Force> changedForces) {
-        // Gather the necessary sending clients; this list may contain null if some units 
+        // Gather the necessary sending clients; this list may contain null if some units
         // cannot be affected at all, i.e. are enemies to localplayer and all his bots
         Set<Client> senders = new HashSet<>();
         senders.addAll(changedEntities.stream().map(this::correctSender).distinct().collect(toList()));
         senders.addAll(changedForces.stream().map(this::correctSender).distinct().collect(toList()));
-        
+
         for (Client sender: senders) {
             if (sender == null) {
                 continue;
             }
             List<Entity> enList = changedEntities.stream().filter(e -> correctSender(e).equals(sender)).collect(toList());
             List<Force> foList = changedForces.stream().filter(f -> correctSender(f).equals(sender)).collect(toList());
-            
+
             if (foList.isEmpty()) {
-                sender.sendUpdateEntity(enList);   
+                sender.sendUpdateEntity(enList);
             } else {
                 sender.sendUpdateForce(foList, enList);
             }
         }
     }
-    
+
     void sendSingleUpdate(Collection<Entity> changedEntities, Collection<Force> changedForces) {
         if (!areAllied(changedEntities, changedForces)) {
             LogManager.getLogger().error("Cannot send force update unless all changed entities and forces are allied!");
             return;
         }
-        
+
     }
 
-    /** 
+    /**
      * Returns the best sending client for an update of the given entity or
      * null if none can be found (entity is an enemy to the local player and all his bots)
      */
@@ -1141,22 +1141,22 @@ public class LobbyActions {
         Player owner = entity.getOwner();
         if (localPlayer().equals(owner)) {
             return client();
-        } else if (client().bots.containsKey(owner.getName())) {
-            return client().bots.get(owner.getName());
+        } else if (client().localBots.containsKey(owner.getName())) {
+            return client().localBots.get(owner.getName());
         } else if (!localPlayer().isEnemyOf(owner)) {
             return client();
         } else {
-            for (Client bot: client().bots.values()) {
+            for (Client bot: client().localBots.values()) {
                 if (!bot.getLocalPlayer().isEnemyOf(owner)) {
                     return bot;
                 }
             }
         }
-        
+
         return null;
     }
-    
-    /** 
+
+    /**
      * Returns the best sending client for an update of the given force or
      * null if none can be found (force is an enemy to the local player and all his bots)
      */
@@ -1164,12 +1164,12 @@ public class LobbyActions {
         Player owner = game().getForces().getOwner(force);
         if (localPlayer().equals(owner)) {
             return client();
-        } else if (client().bots.containsKey(owner.getName())) {
-            return client().bots.get(owner.getName());
+        } else if (client().localBots.containsKey(owner.getName())) {
+            return client().localBots.get(owner.getName());
         } else if (!localPlayer().isEnemyOf(owner) || isEditable(force)) {
             return client();
         } else {
-            for (Client bot: client().bots.values()) {
+            for (Client bot: client().localBots.values()) {
                 if (!bot.getLocalPlayer().isEnemyOf(owner)) {
                     return bot;
                 }
@@ -1178,25 +1178,25 @@ public class LobbyActions {
         return null;
     }
 
-    /** 
+    /**
      * Returns true when the given entity may be configured by the local player,
      * i.e. if it is his own unit or one of his bot's units.
      * <P>Note that this is more restrictive than the Server is. The Server
-     * accepts entity changes also for teammates so that entity updates that 
-     * signal transporting a teammate's unit don't get rejected. 
+     * accepts entity changes also for teammates so that entity updates that
+     * signal transporting a teammate's unit don't get rejected.
      * I think it's important to generally limit entity changes by other players
      * to avoid collisions of updates.
-     * TODO: A possible enhancement might be a GM mode for MM, where only one 
+     * TODO: A possible enhancement might be a GM mode for MM, where only one
      * player is allowed to change everything.
      */
     boolean isEditable(Entity entity) {
-        return client().bots.containsKey(entity.getOwner().getName())
+        return client().localBots.containsKey(entity.getOwner().getName())
                 || (entity.getOwnerId() == localPlayer().getId())
                 || (entity.partOfForce() && isSelfOrLocalBot(game().getForces().getOwner(entity.getForceId())))
                 || (entity.partOfForce() && isEditable(game().getForces().getForce(entity)));
     }
 
-    /** 
+    /**
      * Returns true when the given entity may NOT be configured by the local player,
      * i.e. if it is not own unit or one of his bot's units.
      * @see #isEditable(Entity)
@@ -1205,7 +1205,7 @@ public class LobbyActions {
         return !isEditable(entity);
     }
 
-    /** 
+    /**
      * Returns true when all given entities may be configured by the local player,
      * i.e. if they are his own units or one of his bot's units.
      * @see #isEditable(Entity)
@@ -1229,42 +1229,42 @@ public class LobbyActions {
     boolean entityInLocalTeam(Entity entity) {
         return !localPlayer().isEnemyOf(entity.getOwner());
     }
-    
+
     boolean isSelfOrLocalBot(Player player) {
-        return client().bots.containsKey(player.getName()) || localPlayer().equals(player);
+        return client().localBots.containsKey(player.getName()) || localPlayer().equals(player);
     }
 
     /** Returns true if the entity is an enemy of the local player. */
     boolean isLocalEnemy(Entity entity) {
         return localPlayer().isEnemyOf(entity.getOwner());
     }
-    
+
     /**
      * A force is editable to the local player if any forces in its force chain
      * (this includes the force itself) is owned by the local player or one of the
      * local bots. This allows editing forces of other players if they are a subforce
-     * of a local/bot force.  
+     * of a local/bot force.
      */
     boolean isEditable(Force force) {
         List<Force> chain = game().getForces().forceChain(force);
         return chain.stream().map(f -> game().getForces().getOwner(f)).anyMatch(this::isSelfOrLocalBot);
     }
-    
+
     boolean isEditable(int forceId) {
-        return game().getForces().contains(forceId) && isEditable(game().getForces().getForce(forceId)); 
+        return game().getForces().contains(forceId) && isEditable(game().getForces().getForce(forceId));
     }
-    
+
     boolean areForcesEditable(Collection<Force> forces) {
         return forces.stream().allMatch(this::isEditable);
     }
-    
+
     /**
      * Returns true if no two of the given entities are enemies. This is
-     * true when all entities belong to a single player. If they belong to 
-     * different players, it is true when all belong to the same team and 
+     * true when all entities belong to a single player. If they belong to
+     * different players, it is true when all belong to the same team and
      * that team is one of Teams 1 through 5 (not "No Team").
      * <P>Returns true when entities is empty or has only one entity. The case of
-     * entities being empty should be considered by the caller.  
+     * entities being empty should be considered by the caller.
      */
     private boolean areAllied(Collection<Entity> entities) {
         if (entities.isEmpty()) {
@@ -1274,9 +1274,9 @@ public class LobbyActions {
         Entity randomEntry = entities.stream().findAny().get();
         return entities.stream().noneMatch(e -> e.isEnemyOf(randomEntry));
     }
-    
+
     /**
-     * Returns true if no two of the given entities and forces are enemies. Also checks 
+     * Returns true if no two of the given entities and forces are enemies. Also checks
      * between forces and entities.
      * @see #areAllied(Collection)
      * @see #areForcesAllied(Collection)
@@ -1297,16 +1297,16 @@ public class LobbyActions {
         Force randomForce = forces.stream().findAny().get();
         Player forceOwner = game().getForces().getOwner(randomForce);
         return areAllied(entities) && areForcesAllied(forces) && !entityOwner.isEnemyOf(forceOwner);
-        
+
     }
-    
+
     /**
      * Returns true if no two of the given forces are enemies. This is
-     * true when all forces belong to a single player. If they belong to 
-     * different players, it is true when all belong to the same team and 
+     * true when all forces belong to a single player. If they belong to
+     * different players, it is true when all belong to the same team and
      * that team is one of Teams 1 through 5 (not "No Team").
      * <P>Returns true when forces is empty or has only one force. The case of
-     * forces being empty should be considered by the caller.  
+     * forces being empty should be considered by the caller.
      */
     private boolean areForcesAllied(Collection<Force> forces) {
         if (forces.isEmpty()) {
@@ -1317,19 +1317,19 @@ public class LobbyActions {
         Player owner = game().getForces().getOwner(randomEntry);
         return forces.stream().noneMatch(f -> game().getForces().getOwner(f).isEnemyOf(owner));
     }
-    
+
     private Game game() {
         return lobby.game();
     }
-    
+
     private Client client() {
         return lobby.client();
     }
-    
+
     private JFrame frame() {
         return lobby.getClientgui().getFrame();
     }
-    
+
     private Player localPlayer() {
         return client().getLocalPlayer();
     }
