+++ megamek/src/megamek/client/ui/swing/lobby/PlayerTablePopup.java
@@ -15,7 +15,7 @@
  *
  * You should have received a copy of the GNU General Public License
  * along with MegaMek. If not, see <http://www.gnu.org/licenses/>.
- */ 
+ */
 package megamek.client.ui.swing.lobby;
 
 import java.awt.event.ActionListener;
@@ -31,14 +31,14 @@ import megamek.common.util.CollectionUtil;
 
 import static megamek.client.ui.swing.util.UIUtil.*;
 
-/** 
+/**
  * A popup menu for the lobby's player table.
  * Offers configuration, bot settings and team assignment.
- * 
+ *
  * @author Simon (Juliez)
  */
 class PlayerTablePopup {
-    
+
     static final String PTP_CONFIG = "CONFIG";
     static final String PTP_BOTREMOVE = "BOTREMOVE";
     static final String PTP_BOTSETTINGS = "BOTSETTINGS";
@@ -46,30 +46,30 @@ class PlayerTablePopup {
     static final String PTP_DEPLOY = "DEPLOY";
     static final String PTP_REPLACE = "REPLACE";
 
-    static ScalingPopup playerTablePopup(ClientGUI clientGui, ActionListener listener, 
+    static ScalingPopup playerTablePopup(ClientGUI clientGui, ActionListener listener,
             Collection<Player> players) {
-        
+
         ScalingPopup popup = new ScalingPopup();
-        
+
         var cl = clientGui.getClient();
         var isOnePlayer = players.size() == 1;
         var singlePlayer = CollectionUtil.anyOneElement(players);
         var allOwnedBots = players.stream().allMatch(cl::isLocalBot);
-        var isConfigurable = isOnePlayer 
+        var isConfigurable = isOnePlayer
                 && (allOwnedBots || (cl.getLocalPlayer().equals(singlePlayer)));
         var allConfigurable = players.stream().allMatch(p -> cl.isLocalBot(p) || cl.getLocalPlayer().equals(p));
         var isSingleGhost = isOnePlayer && singlePlayer.isGhost();
-        
+
         popup.add(menuItem("Player Settings...", PTP_CONFIG, isConfigurable, listener));
         popup.add(teamMenu(allConfigurable, listener));
         popup.add(startPosMenu(allConfigurable, listener));
         popup.add(ScalingPopup.spacer());
         popup.add(menuItem("Remove Bot", PTP_BOTREMOVE, isOnePlayer && allOwnedBots, listener));
         popup.add(menuItem("Bot Settings...", PTP_BOTSETTINGS, isOnePlayer && allOwnedBots, listener));
-        popup.add(menuItem("Replace Player...", PTP_REPLACE, isSingleGhost, listener));
-        
+        popup.add(menuItem("Edit Bots...", PTP_REPLACE, isSingleGhost, listener));
+
         return popup;
-        
+
     }
 
     /** Returns the "Team" submenu, allowing to assign a player to a team. */
@@ -82,7 +82,7 @@ class PlayerTablePopup {
         menu.setEnabled(enabled);
         return menu;
     }
-    
+
     /** Returns the "Starting Position" submenu, allowing to assign deployment positions. */
     private static JMenu startPosMenu(boolean enabled, ActionListener listener) {
         JMenu menu = new JMenu("Deployment Area");
